nidl/estimators: split VAE estimators and backbones.
